+++ src/XgpLib.SyncService.Application/DTOs/ReceiveMessagesRequest.cs
@@ -5,4 +5,4 @@
 /// </summary>
 /// <param name="Queue">The queue from which messages will be received.</param>
 /// <param name="MaxMessages">The maximum number of messages to receive (default: 10).</param>
-public record ReceiveMessagesRequest(string Queue, int MaxMessages = 10);
+public record ReceiveMessagesRequest(string Queue, short MaxMessages = 10);
